efficiency

+++ hier/getProbMapTopDown.m
@@ -16,7 +16,7 @@ function [res,resInds] = getProbMapTopDown(cellMapStruct,cellParams,ruleInd,slot
     centreLoc = centre2CellFrame(centre(1:2),strides(1:2),origin(1:2));
     refLoc = centre2CellFrame(cellMapStruct.refPoints(:,ruleInd,slot)',strides(1:2),origin(1:2));
     
-    idx = bsxfun(@plus,locs,[centreLoc-refLoc,0]);
+    idx = int32(bsxfun(@plus,locs,[centreLoc-refLoc,0]));
     
     badInds = find(idx(:,1) < 1 | ...
                    idx(:,2) < 1 | ...
